[buteo-sync-plugins-social] Sync VK images from zero-timestamp albums. Contributes to JB#33230
This commit ensures that we sync photos from VK albums which have a
created/updated timestamp of zero - as these albums are often "system"
albums (like Profile Pictures, Wall Photos, etc).

In that case, we cannot rely on the timestamp difference allowing us
to reduce the delta, and so we need to request all photos from the
album and then calculate the delta manually.

Contributes to JB#33230