[buteo-sync-plugins-social] Set VK calendar entry address if available. Contributes to JB#33268
This commit ensures that we set the location of the event to the
value of the address field of the place contained within the JSON
response.

Note that this will not be set if the owner instead set a place by
id or by latitude/longitude, but those require an extra request to
pull information for, so for now we just support address locations.

Contributes to JB#33268