[buteo-sync-plugins-social] Map locality to gd:city field. Contributes to MER#1555
Previously, there were two fields which we mapped to the QtContacts
locality field on downsync: gd:city and gd:neighborhood.
On upsync, however, we always mapped the locality field to the
gd:neighborhood field, which resulted in great confusion.

This commit ensures that the locality field is a direct mapping
of the gd:city field.  This means that we ignore the gd:neighborhood
altogether, but should result in less confusion since gd:city is
the more commonly provided data.

Contributes to MER#1555