[buteo-sync-plugins-social] Fix upsync of Google Calendar event removal. Contributes to JB#36033
In some cases (where the last-sync timestamp falls within 999 msecs of
when a downsynced event is stored to the database) a local deletion
of that event wouldn't have been upsynced appropriately, due to mkcal
deletedIncidences() semantics.  This commit works around that issue
by ensuring that we retrieve all such locally-deleted events which
were created within the 1-second timeslice around the sync timestamp.

Contributes to JB#36033