[buteo-sync-plugins-social] Early-out finalize() for Google Contacts in error case. Contributes to JB#38210
Previously, a bug in the finalize() method meant that even if an error
occurred during synchronization, we would still attempt finalize the
sync cycle (including potentially destructive operations such as
transforming the contact avatars stored in the database).

This commit ensures that we treat an earlier error as if the sync
was aborted.

Contributes to JB#38210