[buteo-sync-plugins-social] Use SyncToken for Google Calendar delta sync. Contributes to JB#44316
Previously, we used updatedMin timestamp "since" anchor to fetch
changes since last sync.  This commit updates the plugin so that
it uses the "sync token" which Google provides, so that we can
more accurately fetch changes since the last sync, and avoid
the "updatedMin too far in the past" problem which can be hit in
certain circumstances.

We store the sync tokens as account settings, as mkcal doesn't
allow arbitrary metadata to be stored for notebooks.