[buteo-sync-plugins-social] Fix Google calendar event duplication. Contributes to JB#47783
This commit fixes several issues which together could lead to errors
which result in event duplication in some cases.

- sync anchor timestamp is now stored exclusively in the notebook
  as the syncDate property, to avoid atomicity issues

- sync anchor token is now stored exclusively in the notebook
  as a custom property (instead of account setting) to avoid
  atomicity issues

- detection of server-side sync token invalidation has been
  improved, allowing the sync envelope to be tweaked appropriately

- detection of events copied between calendars via Google web UI
  has been added, to ensure we don't attempt to store multiple
  events with the same UID (which could trigger a clean-sync cycle).

It also bumps the plugin version to ensure that a clean sync will
be triggered on the first run, to clean up any duplications which
may have previously occurred.

Contributes to JB#47783