Ignore 403 ContentOperationNotPermitted errors due to shared calendars
Some changes (e.g. changing the organizer of an event in a shared
calendar) are not allowed, even if the user has Owner access to the
shared calendar (since the organizer field is managed specially
by Google in the shared-calendar case, allowing only the organizer
or the creator of the event to write changes there).

If we hit such a failure, ignore it and allow sync to succeed.