[buteo-sync-plugins-social] Discard spurious local modification due to remote organizer change. Contributes to JB#47783
If a local modification is reported as a change since the last sync,
but the data in the event is identical to the data reported by the
server for that event except for the organizer email address,
then it is probably a spurious change (since Google will add the
organizer email automatically).

This commit detects that case and doesn't attempt to upsync the
modification.