[buteo-sync-plugins-social] Discard spurious local addition due to downsync. Contributes to JB#47783
Previously, we would determine that local additions caused by a
downsync during the previous sync cycle should be treated as
local modifications, and upsynced.

This commit checks whether a local modification has indeed occurred
by checking the lastModified date stamp, and if no modification has
occurred, discarding the change to prevent spurious upsync.

It also fixes a bug where the wrong "since" date time would be used
when determining the local changeset, as it may be using one from
a different notebook by mistake.  This commit updates the code so
that the appropriate previous sync date for the specific notebook
is used.