[buteo-sync-plugins-social] Don't pass oauth token as key query parameter. Contributes to JB#50140
Google calendar sync recently started (intermittently) failing.
I believe the reason is that for some API endpoints, Google now
requires an API specific key to be passed as a query parameter,
aside from the oauth2 token which is passed in the authorization
header.  Even though the Calendar API is not one of the APIs which
requires this type of API key, passing the oauth2 token as a key
parameter was causing issues.

This commit ensures we don't pass the token as a query parameter.
It is already provided in the authorization header.