[buteo-sync-plugins-social] Fix bugs in Google Contact sync. Contributes to JB#51541
We no longer hardcode the name of the default My Contacts collection,
so when searching for it we cannot look for the hardcoded name.
Instead, mark the collection with a special piece of metadata.

Also fix a bug in etag detection.

Finally, ensure we ignore detail access constraints when updating
detail data during sync.