[buteo-sync-plugins-social] Use gd:fullName in Google Contacts API. JB#51541
Encode and decode the gd:fullName entry in the atom feed. The server
always sends a fullName composed from the first name, last name,
prefix etc. so encoding this in the request will avoid a spurious
modification from the server on the next sync after it reassembles the
fullName on the server side.