[buteo-sync-plugins-social] Fix Google contact sync issues. JB#51541
When upsyncing contacts, the server will reply with an updated etag
for each contact in postFinishedHandler(). Save the etags to their
contacts before calling localChangesStoredRemotely(), which will save
the updated contacts to the database. Also save the avatars at this
point instead of in finalize(), as calling QContactManager::saveContacts()
will flag the contacts as modified and cause spurious modification
detections on the next upsync.

Also:
- Ensure the correct detail is looked up for the etag data
- Don't remove contacts from the localAdded/Modified lists in
batchRemoteChanges(), because these lists are then passed onto
localChangesStoredRemotely() for saving
- Save the last sync time in finalize() so that it is not saved if the
sync failed.