[buteo-sync-plugins-social] Use client timestamp for sync time. Contributes to JB#51507
The deleted and modified times of events are stored using the client
timestamp. If the timestamp returned by the server is used to determine
the upload delta, this can result in the wrong entries being uploaded.
In particular it was resulting in events deleted on the server being
uploaded as new deletions to the server, causing an error.

This change sets the notebook last sync time using the current client
timestamp, to ensure only entries changed after the latest sync are
uploaded to the server.

In the reverse direction the server uses a sync token to determine which
events to sync, so the change doesn't affect which events are
downloaded.