[buteo-sync-plugins-social] Don't set Google gd:fullName as customLabel field. Fixes JB#51831
The customLabel field is used as the default display label in the
QtContacts backend, and since the customLabel for a Google contact is
only updated during a sync operation, the customLabel (and hence the
contact's display label) was not updated when the contact name changed
outside of a sync operation.

It's not necessary to set this as the customLabel field in any case,
because we need to always send the latest version of the formatted name
(according to any local name changes for the contact) in order to avoid
spurious modification notifications from the server.