[buteo-sync-plugins-social] Record sync time as time started, rather than time ended. Contributes to JB#51507
This change records the time at the start of the sync and uses this as
the previous sync time. This ensures that any events modified after the
start of the sync will be uploaded in the next sync.

Deleted events (both local and remote, where the change is already on
the server, are also now purged from the database after a successful
sync to avoid any attempt to upload the deletion to the server on the
next sync (which would cause an error).