[buteo-sync-plugins-social] Translate between mkcal and Google EXDATEs. Contributes to JB#51507
Whenever an event from a recurring series is given an exception event
(e.g. a single event in the series is edited), a new child event is
created for the exception.

mkcal also adds an EXDATE to the recurrence rule for the parent to
remove the original entry that was edited. In contrast, Google doesn't
add an EXDATE, but just automatically performs the removal.

This causes problems for syncing.

This change adds the extra EXDATEs to the recurrence rules of downsynced
events as needed by mkcal, and removes them from the recurrence rules of
upsynced events.