[buteo-sync-plugins-social] Set recurrence start time when converting from json. Contributes to JB#51507
The recurrence rules store the event start time separately from the
event itself. If set incorrectly (or not at all) the rules end up being
badly formed, and other actions (e.g. dissociating event exceptions)
could fail.

Events extracted from the database have recurrence start times
automatically set to the event start time, but this wasn't happening
during json deserialisation. This change adds it in.

The issue was triggered when an update for the main event was received
from google (resulting in a json deserialisation), followed in the same
sync by a new recurrence exception for the same event. In this case, the
dissociation for the recurrence was failing due to the lack of start
time in the recurrence rule. If the changes were received in separate
syncs the problem didn't arise, since the main event had been "washed"
through the database.