[buteo-sync-plugins-social] Keep track of additions for later modifications in the same sync. Contributes to JB#51507
If an event is added during a sync, its details must be kept track of in
case further modifications occur to the event (e.g. exceptions to or
deletions from the recurrence rule) later in the same sync.

The additions were already being carefully ordered to ensure they're
inserted first, but there were some details not added to the maps for
finding the parent event later. This change inserts the event details in
these maps as the sync progresses so that later modifications can be
applied successfully.