[buteo-sync-plugins-social] Pass recurrenceId to the network reply as a QVariant. Contributes to JB#52657
Various items of data are attached to the QNetworkReply for use when the
reply returns. The recurrenceId was being parsed in as Qt::ISODate
format, but written out as Qt::TextDate format. As a result, the
recurrenceId was always empty, and the parent recurring event was being
amended by the upsync response rather than the exception.

This change attaches the recurrenceId property to the QNetworkReply
object as a QVariant, so that it's restored correctly.