[buteo-sync-plugins-social] Sequence exception insertions after their parents. Contributes to JB#52657
For Google calender synchronisation, if an exception to a recurring
event is upsynced as an insertion to the remote calendar at the same
time as its parent recurring event, the two have to be sequenced
carefully. The upsync of the parent has to complete fully before the
child can be uploaded, otherwise Google won't find the parent and
returns an error.

This change adds a sequencing step that allows the relevant insertion
upsyncs to be sequenced in this way.