[buteo-sync-plugins-social] Avoid performing multiple clean-syncs. Contributes to JB#52608
When an account is initalised, a clean sync is initiated. However,
because the sync plugin version was left unset, the second was also
performed as a clean sync, even if the first was successful.

This change ensures the protocol version number gets stored correctly
after the first full sync in order to avoid this.