[buteo-sync-plugins-social] Differentiate between 403 errors. Contributes to JB#52608
If Google returns a 403 error it can mean modifying a non-authored
shared event, but it can also mean a usage limit has been reached. The
first is ignorable for the full sync, whereas the second should block
the full sync. They therefore need to be distinguished and handled
differently.

Delete events that return 404 errors are also handled differently. These
are ignored rather than causing the sync to fail.