[buteo-sync-plugins-social] Add dummy parent if Google sends an exception without one. Contributes to JB#52608
If Google sends an exception event during a sync for which the parent
recurring event doesn't exist, this would previously cause the entire
sync to fail. This change instead creates a dummy parent event, matching
the exception, which the exeption is then dissociated from.

This allows the sync to succeed and the event to appear. This also
matches the behaviour of the caldav sync plugin.