[buteo-sync-plugins-social] Clamp min sync time to within a year. Contributes to JB#52608
For accounts that have never experienced a successful sync, the last
sync time may return 0 with an invalid sync token. This will cause the
next sync to be clean, but with the requested minimum time set to 0
(i.e. "1970-01-01T00:00:00Z").

In practice, it never makes sense to sync back this far, so now the
start time is bounded below by a year prior to the current time.