[buteo-sync-plugins-social] Ensure clean sync is applied in full. Contributes to JB#52608
The google sync plugin can be tricked into performing a "partial clean
sync" if both the syncToken and syncTime are invalid. This state can
occur, for example, if the account has never managed to complete a fully
successful sync.

The result of this state is that all events are considered as insertions
(as opposed to modifications), while at the same time the notebook isn't
recreated from scratch. As a result, duplicate entries are added to the
calendar on each sync.

This change fixes this by identifying the case where the syncToken and
syncTime are both invalid and ensuring the clean sync semantics are
applied throughout the sync (so everything is an insertion, but the
notebook is also recreated and so starts empty).