[buteo-sync-plugins-social] Update Contacts plugin to use Google People API. JB#50799
The Google Contacts API is deprecated and the service will cease to
work in June 2021. Update the contacts plugin to use the newer People
API instead.

Main changes:
- Use JSON instead of XML to format request and response data.
- Make multi-part HTTP requests for batched upsync requests.
- Upsync avatar changes separately from other contact data as per the
People API requirements.
- Always save both the remote url and local file path for an avatar
instead of transforming the remote url to a local file after downloading.
- If a downloaded avatar's file extension does not match its meta type
(e.g. is png but has jpg extension) rename the file to match the type.