[buteo-sync-plugins-social] Ignore Google contact date if it cannot be read. JB#53422
The date of a Google birthday or event may be missing a 'year' value
even if it is otherwise valid. Also, a birthday may be just some
freeform text without a date. In these cases, ignore the value,
otherwise we will upload an invalid date on the next upsync and
overwrite the value on the server.