[buteo-sync-plugins-social] Disable updating last modified time on event save. Contributes to JB#53475
Previously, the backend would automatically update the last modified
time for an event when the event was saved.  Sync plugins (which
save updates to events due to server-side changes) should avoid this.

Recently, API was added to the backend calendar API to allow this
behaviour to be disabled by sync plugins.  This commit takes this
API into use, and removes the old work-arounds.